requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0

maritalk==0.2.7
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.97.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
tiktoken==0.7.0
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3

supabase==2.4.3
PyJWT==2.8.0
